+++ CHANGELOG.md
@@ -272,6 +272,8 @@ Add new OpenaAI models `gpt-4o` and `gpt-4o-2024-05-13`
 -   List OpenAI models dynamically
 -   Export types `LlmTemplateJson`, `SimpleTemplateJson`, `ScriptJson`, `PromptDialogJson` from `@promptbook/types`
 -   `ModelRequirements.modelName` is not required anymore
+-   `PromptbookExecutor` does not require `onProgress` anymore
+-   `ExecutionTools` does not require `userInterface` anymore, when not set, the user interface is disabled and promptbook which requires user interaction will fail
 
 ## In prepare